Moved inode stuff to its own file
