+++ frontend/src/save-formats/N64/Components/Basics.js
@@ -6,4 +6,6 @@ export default class N64Basics {
   static NUM_PAGES = 128;
 
   static PAGE_SIZE = 256;
+
+  static FIRST_SAVE_DATA_PAGE = 5;
 }
